requirements.txt
Flask==2.2.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
